+++ lib/phoenix_storybook/live/story/component_iframe_live.ex
@@ -58,8 +58,12 @@ defmodule PhoenixStorybook.Story.ComponentIframeLive do
 
   defp assign_color_mode_class(socket = %{assigns: assigns}) do
     class =
-      if assigns.color_mode == "dark" do
-        assigns.backend_module.config(:color_mode_sandbox_dark_class, "dark")
+      if assigns.backend_module.config(:color_mode) do
+        case assigns.color_mode do
+          "dark" -> assigns.backend_module.config(:color_mode_sandbox_dark_class, "dark")
+          "light" -> assigns.backend_module.config(:color_mode_sandbox_light_class)
+          _ -> nil
+        end
       end
 
     assign(socket, :color_mode_class, class)
@@ -188,4 +192,14 @@ defmodule PhoenixStorybook.Story.ComponentIframeLive do
   end
 
   def handle_event(_, _, socket), do: {:noreply, socket}
+
+  def handle_info(message, socket) do
+    story = socket.assigns[:story]
+
+    if is_atom(story) and function_exported?(story, :handle_info, 2) do
+      story.handle_info(message, socket)
+    else
+      {:noreply, socket}
+    end
+  end
 end