+++ CHANGELOG.md
@@ -9,9 +9,12 @@
 > - You can also [buy me a coffee](https://ko-fi.com/C0C316P9FN), as it would encourage me spending
 >   more of my free time to fixing bugs and developing new features 🤗
 
-## v0.9.4 (not yet released)
+## v0.9.4 (not released)
+- **feature**: [support custom light class in color mode sandbox](https://github.com/phenixdigital/phoenix_storybook/issues/722)
+- **feature**: allow live_component stories to implement `handle_info/2` for parent message handling (e.g. `send_update/2`)
 - **improvement**: [allow overriding FontAwesome CSS path to use app-provided assets](https://github.com/phenixdigital/phoenix_storybook/issues/739)
 - **bugfix**: [storybook asset hash query param renamed to avoid live reload URL rewriting](https://github.com/phenixdigital/phoenix_storybook/issues/699)
+- **bugfix**: [Error when using hooks defined as a subclass of LiveView's v1.1 new ViewHook](https://github.com/phenixdigital/phoenix_storybook/issues/663)
 
 ## v0.9.3 (2025-09-15)
 - **feature**: [accompany variations with optional notes](https://github.com/phenixdigital/phoenix_storybook/issues/571) - 🙏 [@xxdavid](https://github.com/xxdavid)